refactor(attendance): Add try catch on autoCheckout function service, Add unit test for autoCheckout also rename Mood constant to Enum
