+++ src/models/attendances/enums/mood.enums.ts
